+++ lib/trolley/Balance.rb
@@ -1,4 +1,4 @@
-module PaymentRails
+module Trolley
   class Balance
   end
 end
