+++ lib/trolley/Client.rb
@@ -6,7 +6,7 @@ require 'uri'
 require 'json'
 require "rubygems"
 
-module PaymentRails
+module Trolley
   class Client
     def initialize(config)
       @config = config
@@ -47,7 +47,7 @@ module PaymentRails
                 'Authorization': generate_authorization(time, endPoint, method, body),
                 'Content-Type': 'application/json',
                 'Trolley-Source': "ruby-sdk_#{spec.version}"}
-                
+
       if method === "GET"
         request = Net::HTTP::Get.new(uri.request_uri, headers)
       elsif method === "POST"
