+++ lib/trolley/Invoice.rb
@@ -1,4 +1,4 @@
-module PaymentRails
+module Trolley
   class Invoice
     attr_accessor(
       :id,
