+++ lib/trolley/gateways/InvoicePaymentGateway.rb
@@ -1,7 +1,7 @@
 require_relative '../Client.rb'
 require_relative 'GatewayHelper'
 
-module PaymentRails
+module Trolley
   class InvoicePaymentGateway
     include GatewayHelper
