+++ spec/integration/RecipientTest.rb
@@ -1,14 +1,7 @@
 require_relative 'helper'
 
 class RecipientTest < Test::Unit::TestCase
-  def setup
-    @client = PaymentRails.client(
-      ENV.fetch('SANDBOX_API_KEY'),
-      ENV.fetch('SANDBOX_SECRET_KEY'),
-      'production',
-      proxy_uri: ENV['PROXY_URI']
-    )
-  end
+  include ApiClientHelper
 
   def test_create
     uuid = SecureRandom.uuid.to_s
