+++ spec/integration/helper.rb
@@ -1,4 +1,15 @@
+require_relative '../../lib/trolley'
 require 'dotenv/load'
-require_relative '../../lib/paymentrails'
 require 'test/unit'
 require 'securerandom'
+
+module ApiClientHelper
+  def setup
+    @client = Trolley.client(
+      ENV.fetch('SANDBOX_API_KEY'),
+      ENV.fetch('SANDBOX_SECRET_KEY'),
+      'production',
+      proxy_uri: ENV['PROXY_URI']
+    )
+  end
+end
